requirements.txt
opencv_python==4.5.2.54
matplotlib==3.4.2
Flask==2.0.1
numpy==1.19.5
protobuf==3.17.3
gunicorn==20.1.0
google-cloud==0.34.0
google-cloud-vision==2.3.2
